pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "music-of-the-day"
version = "0.1.0"
description = "Daily news → semantic meaning → piano music"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }

authors = [
  { name = "CentralBrancher", email = "centralbrancher@gmail.com" }
]

dependencies = [
  "torch>=2.0",
  "transformers>=4.35",
  "sentence-transformers>=2.2",
  "numpy>=1.24",
  "scikit-learn>=1.3",
  "scipy>=1.10",
  "pretty_midi>=0.2.10",
  "music21>=8.3",
  "tqdm>=4.66",
  "pyyaml>=6.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "black>=23.9",
  "ruff>=0.1.6"
]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.ruff]
select = ["E", "F", "I"]
line-length = 88
